day15_capstone/aim2_research/requirements.txt
git+https://github.com/hypercycle-development/pyhypercycle-aim.git#egg=pyhypercycle_aim
groq==0.9.0
httpx==0.27.0
google-search-results==2.4.2
pydantic==2.6.0
python-dotenv==1.0.0